Add new "Pricing" section
The new section (and the according links in the header and footer)
are shown when pricing/pricing.md exists and is set to "draft: false".
Example data is available in exampleSite/data.

At the moment, this part does not use bootstrap. As soon as BS v4 is
available and integrated, this can be changed to use the new flexbox
feature in BS v4. A separate pricing.css is used to make this update
easier.